// packages/discord/src/operations/getActiveGuildThreads.ts
export const GetActiveGuildThreadsInput =
  /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
    guild_id: Schema.String.pipe(T.PathParam()),
  }).pipe(T.Http({ method: "GET", path: "/guilds/{guild_id}/threads/active" }));
export type GetActiveGuildThreadsInput = typeof GetActiveGuildThreadsInput.Type;
// ...
export type GetActiveGuildThreadsOutput =
  typeof GetActiveGuildThreadsOutput.Type;

// The operation
export const getActiveGuildThreads = /*@__PURE__*/ /*#__PURE__*/ API.make(
  () => ({
    inputSchema: GetActiveGuildThreadsInput,
    outputSchema: GetActiveGuildThreadsOutput,
    errors: [BadRequest, Forbidden, NotFound] as const,
  }),
);